Remove Collapsible and refactor
