--- src/UI/Accordion/index.js
@@ -1,3 +1,0 @@
-import Collapsible from "./Collapsible";
-
-export default Collapsible;
