--- src/UI/Accordion/styles.scss
@@ -1,25 +1,0 @@
-@import "./../colors";
-
-.accordion_wrapper {
-    border: 1px solid $primaryColor;
-    margin: 20px;
-    min-width: 70vw;
-}
-
-.accordion_title * {
-    display: inline-flex;
-    vertical-align: middle;
-}
-
-.accordion_title {
-    width: 100%;
-    padding: 20px;
-    cursor: pointer;
-}
-
-.accordion_inner {
-    height: 0px;
-    overflow: hidden;
-
-    transition: height 400ms ease;
-}
