+++ src/UI/PluginPreview/styles.scss
@@ -1,5 +1,3 @@
-@import "./../../colors";
-
 * {
     margin: 0;
     padding: 0;
