+++ src/UI/index.js
@@ -1,14 +1,12 @@
-import Collapsible from './Accordion';
-import PluginPreview from './Accordion/PluginPreview'
 import AppBar from './AppBar';
 import Button from './Button';
 import Modal  from './Modal';
 import { MenuItem } from './Menu';
 import Paper from './Paper';
+import PluginPreview from './PluginPreview';
 import Divider from './Divider';
 
 export {
-    Collapsible,
     AppBar,
     Button,
     Modal,
